Add more source dependency test coverage.
