--- subprojects/version-control/src/integTest/groovy/org/gradle/vcs/internal/AbstractVcsIntegrationTest.groovy
@@ -1,51 +1,0 @@
-/*
- * Copyright 2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.vcs.internal
-
-import org.gradle.integtests.fixtures.AbstractIntegrationSpec
-import org.gradle.integtests.fixtures.build.BuildTestFile
-
-abstract class AbstractVcsIntegrationTest extends AbstractIntegrationSpec implements SourceDependencies {
-    BuildTestFile depProject
-
-    def setup() {
-        buildFile << """
-            apply plugin: 'java'
-            group = 'org.gradle'
-            version = '2.0'
-            
-            dependencies {
-                compile "org.test:dep:latest.integration"
-            }
-        """
-        file("src/main/java/Main.java") << """
-            public class Main {
-                Dep dep = null;
-            }
-        """
-        buildTestFixture.withBuildInSubDir()
-        depProject = singleProjectBuild("dep") {
-            buildFile << """
-                allprojects {
-                    apply plugin: 'java'
-                    group = 'org.test'
-                }
-            """
-            file("src/main/java/Dep.java") << "public class Dep {}"
-        }
-    }
-}
