+++ subprojects/version-control/src/integTest/groovy/org/gradle/vcs/internal/NestedSourceDependencyIntegrationTest.groovy
@@ -21,7 +21,7 @@ import org.gradle.util.TextUtil
 import org.gradle.vcs.fixtures.GitFileRepository
 import org.junit.Rule
 
-class SourceDependencyIntegrationTest extends AbstractIntegrationSpec {
+class NestedSourceDependencyIntegrationTest extends AbstractIntegrationSpec {
     @Rule
     GitFileRepository first = new GitFileRepository('first', testDirectory)
     @Rule
